Move getScores api to separate directory
