+++ src/api/getScores.js
@@ -2,7 +2,7 @@ import { API } from '../constants';
 
 const sorter = (object) => object.sort((a, b) => (a.score > b.score ? -1 : 1));
 
-export const getScores = async () => {
+const getScores = async () => {
   try {
     const res = await fetch(`${API}/games/DgRb9bRvC0KaxNWE7YIg/scores`, {
       headers: {
@@ -19,25 +19,4 @@ export const getScores = async () => {
   }
 };
 
-const updateScores = async ({ user, score }) => {
-  try {
-    const res = await fetch(`${API}/games/DgRb9bRvC0KaxNWE7YIg/scores`, {
-      method: 'POST',
-      headers: {
-        'Content-type': 'application/json',
-        Accept: 'application/json',
-      },
-      body: JSON.stringify({
-        user,
-        score,
-      }),
-    });
-    await res.json();
-
-    return getScores();
-  } catch (error) {
-    throw new Error(error);
-  }
-};
-
-export default updateScores;
+export default getScores;
